requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-core
python-dotenv
pdfplumber
python-docx
pytesseract
pandas
Pillow
python-multipart
groq
PyPDF2
transformers
httpx
faiss-cpu
sentence-transformers
openai
pypdf
easyocr
tiktoken
tenacity
pdf2image